*Minor changes to satisfy cppcheck and Splint warnings. *Altered RS2 export to prevent duplicate single notes from exporting in the event of overlapping arpeggio phrases. *Added logic to remove the possibility of overlapping arpeggio phrases. *Removed unused functions eof_song_tick_to_msec() and eof_song_msec_to_tick(). *Changed eof_get_beat_length() to use the floating point beat positions, since the calling function was using floating point math anyway.  This should improve the accuracy of the auto-adjust logic. *Changed the handling of time signatures so that denominator (beat unit) alters the length of a beat as it does in music theory. *Removed the double float typecast from numbers that can just be written with a decimal point, since that's the type they are considered to be in C.

+++ trunk/src/bf.c
@@ -101,11 +101,11 @@ int eof_export_bandfuse(EOF_SONG * sp, char * fn, unsigned short *user_warned)
 		}
 		if(ctr2 < sp->beats)
 		{	//If a beat with a different tempo was found, that beat is written as the end position of this tempo change
-			(void) snprintf(buffer, sizeof(buffer) - 1, "    <tempo start=\"%lu\" end=\"%lu\" tempo=\"%f\">\n", sp->beat[ctr]->pos, sp->beat[ctr2]->pos, (double)60000000.0 / (double)sp->beat[ctr]->ppqn);
+			(void) snprintf(buffer, sizeof(buffer) - 1, "    <tempo start=\"%lu\" end=\"%lu\" tempo=\"%f\">\n", sp->beat[ctr]->pos, sp->beat[ctr2]->pos, 60000000.0 / (double)sp->beat[ctr]->ppqn);
 		}
 		else
 		{	//No remaining beats had a different tempo, the outer for loop's beat is written as the end position of this tempo change
-			(void) snprintf(buffer, sizeof(buffer) - 1, "    <tempo start=\"%lu\" end=\"%lu\" tempo=\"%f\">\n", sp->beat[ctr]->pos, sp->beat[ctr]->pos, (double)60000000.0 / (double)sp->beat[ctr]->ppqn);
+			(void) snprintf(buffer, sizeof(buffer) - 1, "    <tempo start=\"%lu\" end=\"%lu\" tempo=\"%f\">\n", sp->beat[ctr]->pos, sp->beat[ctr]->pos, 60000000.0 / (double)sp->beat[ctr]->ppqn);
 		}
 		(void) pack_fputs(buffer, fp);
 		ctr = ctr2;	//Advance the beat counter
